+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.0.0
 * Introduced a custom resource to run custom artisan cammands
 * Refactored all recipes to use the new custom resource
+* Introduced another custom resource for laravel artisan queues
 
 ## 0.6.0
 * Added support for `php artisan route cache`